feat: add party profile view with edit and delete
