+++ lib/modules/party/screens/party_profile_screen.dart
@@ -1,12 +1,11 @@
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
-import '../../../widgets/app_button.dart';
-import '../../../widgets/app_text_from_field.dart';
-import '../controllers/party_profile_controller.dart';
 import '../../../models/party.dart';
+import '../../../widgets/app_button.dart';
+import '../../../widgets/app_info_row.dart';
+import '../controllers/party_profile_controller.dart';
+import 'edit_party_screen.dart';
 
 class PartyProfileScreen extends StatelessWidget {
   final Party party;
@@ -22,59 +21,49 @@ class PartyProfileScreen extends StatelessWidget {
       body: Obx(() {
         return Stack(
           children: [
-            SingleChildScrollView(
+            Padding(
               padding: const EdgeInsets.all(16),
               child: Column(
                 spacing: 16,
                 children: [
-                  GestureDetector(
-                    onTap: controller.showImagePicker,
-                    child: Obx(() {
-                      final image = controller.photo.value;
-                      return CircleAvatar(
-                        radius: 50,
-                        backgroundImage: image != null
-                            ? FileImage(File(image.path))
-                            : (party.photoUrl != null
-                                ? NetworkImage(party.photoUrl!) as ImageProvider
-                                : null),
-                        child: image == null && party.photoUrl == null
-                            ? const Icon(Icons.camera_alt, size: 40)
-                            : null,
-                      );
-                    }),
-                  ),
-                  AppTextFromField(
-                    controller: controller.name,
-                    label: 'নাম',
-                    hintText: 'পক্ষের নাম লিখুন',
-                    prefixIcon: Icons.person,
-                  ),
-                  AppTextFromField(
-                    controller: controller.phone,
-                    label: 'মোবাইল',
-                    hintText: 'মোবাইল নম্বর লিখুন',
-                    prefixIcon: Icons.phone,
-                    keyboardType: TextInputType.phone,
-                  ),
-                  AppTextFromField(
-                    controller: controller.address,
-                    label: 'ঠিকানা',
-                    hintText: 'ঠিকানা লিখুন',
-                    prefixIcon: Icons.home,
-                    isMaxLines: 3,
-                  ),
-                  const SizedBox(height: 20),
-                  AppButton(
-                    label: 'আপডেট করুন',
-                    onPressed: controller.enableBtn.value
-                        ? controller.updateParty
+                  CircleAvatar(
+                    radius: 50,
+                    backgroundImage: party.photoUrl != null
+                        ? NetworkImage(party.photoUrl!)
                         : null,
+                    child: party.photoUrl == null
+                        ? const Icon(Icons.person, size: 40)
+                        : null,
+                  ),
+                  appInfoRow('নাম', party.name),
+                  appInfoRow('মোবাইল', party.phone),
+                  appInfoRow('ঠিকানা', party.address),
+                  const Spacer(),
+                  Row(
+                    children: [
+                      Expanded(
+                        child: AppButton(
+                          label: 'এডিট',
+                          onPressed: () {
+                            Get.to(() => EditPartyScreen(party: party));
+                          },
+                        ),
+                      ),
+                      const SizedBox(width: 16),
+                      Expanded(
+                        child: AppButton(
+                          label: 'ডিলিট',
+                          onPressed: controller.isDeleting.value
+                              ? null
+                              : controller.deleteParty,
+                        ),
+                      ),
+                    ],
                   ),
                 ],
               ),
             ),
-            if (controller.isLoading.value)
+            if (controller.isDeleting.value)
               const Center(child: CircularProgressIndicator()),
           ],
         );
